Drop the MauiReactor adapter package: the bridge is app-side code
Nalu ships no component-framework adapters. The Nalu.Maui.Navigation.MauiReactor
project is removed; what remains public in the core is the framework-agnostic
seam (IComponentPageFactory / IComponentPageHandle / UseComponentPageFactory<T>()
and the relaxed AddPage<T>() + [AutoNavigationPage] opt-in).

The MauiReactor bridge is now a documented ~35-line copy-paste class:
- canonical sample: Samples/Nalu.Maui.TestApp/MauiReactorComponentPageFactory.cs,
  registered via UseComponentPageFactory<MauiReactorComponentPageFactory>();
- an identical copy lives in the unit tests, keeping the documented code
  exercised headlessly against the real MauiReactor runtime;
- conceptual_docs/navigation-mauireactor.md carries the full snippet and the
  one-line registration; README/index/CLAUDE.md wording updated accordingly.

All suites stay green (745 unit, 30 generator).
