+++ Nalu.Pack.slnf
@@ -7,7 +7,6 @@
       "Source\\Nalu.Maui.Layouts\\Nalu.Maui.Layouts.csproj",
       "Source\\Nalu.Maui.VirtualScroll\\Nalu.Maui.VirtualScroll.csproj",
       "Source\\Nalu.Maui.Navigation\\Nalu.Maui.Navigation.csproj",
-      "Source\\Nalu.Maui.Navigation.MauiReactor\\Nalu.Maui.Navigation.MauiReactor.csproj",
       "Source\\Nalu.Maui.Navigation.SourceGenerators\\Nalu.Maui.Navigation.SourceGenerators.csproj",
       "Source\\Nalu.Maui.Scaffold\\Nalu.Maui.Scaffold.csproj",
       "Templates\\Nalu.Maui.Templates\\Nalu.Maui.Templates.csproj",
