+++ Nalu.slnx
@@ -52,7 +52,6 @@
     <Project Path="Source/Nalu.Maui.Core/Nalu.Maui.Core.csproj" />
     <Project Path="Source/Nalu.Maui.Layouts/Nalu.Maui.Layouts.csproj" />
     <Project Path="Source/Nalu.Maui.Navigation/Nalu.Maui.Navigation.csproj" />
-    <Project Path="Source/Nalu.Maui.Navigation.MauiReactor/Nalu.Maui.Navigation.MauiReactor.csproj" />
     <Project Path="Source/Nalu.Maui.Navigation.SourceGenerators/Nalu.Maui.Navigation.SourceGenerators.csproj" />
     <Project Path="Source/Nalu.Maui.Scaffold/Nalu.Maui.Scaffold.csproj" />
     <Project Path="Source/Nalu.Maui.Scaffold.SourceGenerators/Nalu.Maui.Scaffold.SourceGenerators.csproj" />
