+++ README.md
@@ -35,7 +35,7 @@ Because the Scaffold *hosts* Nalu navigation, every interaction — tab tap, bac
 ### Key Modules
 
 *   **Navigation** [![Nalu.Maui.Navigation NuGet Package](https://img.shields.io/nuget/v/Nalu.Maui.Navigation.svg)](https://www.nuget.org/packages/Nalu.Maui.Navigation/) [![Nalu.Maui NuGet Package Downloads](https://img.shields.io/nuget/dt/Nalu.Maui.Navigation)](https://www.nuget.org/packages/Nalu.Maui.Navigation/)
-    *   A fluent, type-safe MVVM navigation service supporting relative/absolute navigation, guards, typed intents and a built-in **leak detector**. Page registration is **source-generated** (`AddPages()` — trim/AOT-safe, no reflection), and opt-in [navigation state restoration](conceptual_docs/navigation-restore.md) reopens the app exactly where it was. Works with MAUI `Shell`, `NaluShell`, and the Scaffold. MVVM is optional: plain view-only pages and **MVU component pages** ride the same engine — [MauiReactor](https://github.com/adospace/reactorui-maui) via the `Nalu.Maui.Navigation.MauiReactor` adapter, other component frameworks (Comet, BlazorBindings.Maui) via the public `IComponentPageFactory` extension point.
+    *   A fluent, type-safe MVVM navigation service supporting relative/absolute navigation, guards, typed intents and a built-in **leak detector**. Page registration is **source-generated** (`AddPages()` — trim/AOT-safe, no reflection), and opt-in [navigation state restoration](conceptual_docs/navigation-restore.md) reopens the app exactly where it was. Works with MAUI `Shell`, `NaluShell`, and the Scaffold. MVVM is optional: plain view-only pages and **MVU component pages** ride the same engine — [MauiReactor](https://github.com/adospace/reactorui-maui), Comet or BlazorBindings.Maui plug in through the public `IComponentPageFactory` extension point (a small copy-paste bridge; the MauiReactor one is [in the docs](https://nalu-development.github.io/nalu/navigation-mauireactor.html)).
 *   **VirtualScroll** [![Nalu.Maui.VirtualScroll NuGet Package](https://img.shields.io/nuget/v/Nalu.Maui.VirtualScroll.svg)](https://www.nuget.org/packages/Nalu.Maui.VirtualScroll/) [![Nalu.Maui NuGet Package Downloads](https://img.shields.io/nuget/dt/Nalu.Maui.VirtualScroll)](https://www.nuget.org/packages/Nalu.Maui.VirtualScroll/)
     *   A **fast** alternative to the .NET MAUI `CollectionView`, built directly on native `RecyclerView` (Android) and `UICollectionView` (iOS) with a Java-side hot path that keeps per-frame work off the managed heap — fling through thousands of dynamically-sized rows without a stutter. Sections with headers, horizontal layout, carousel mode, pull-to-refresh, animated `ScrollTo`, and long-press **drag reorder** are all built in.
     *   ⚖️ **Dual Licensed**:
