--- Source/Nalu.Maui.Navigation.MauiReactor/Nalu.Maui.Navigation.MauiReactor.csproj
@@ -1,46 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup Label="Package">
-    <Product>Nalu.Maui</Product>
-    <Description>MauiReactor adapter for Nalu.Maui navigation: navigate to MauiReactor stateful components as first-class Nalu pages — lifecycle hooks, typed intents and leaving guards implemented directly on the component, no page model required.</Description>
-    <PackageTags>MAUI;navigation;mvu;mauireactor;components;lifecycle</PackageTags>
-    <RootNamespace>Nalu</RootNamespace>
-    <IsAotCompatible>true</IsAotCompatible>
-    <ProduceReferenceAssembly>true</ProduceReferenceAssembly>
-  </PropertyGroup>
-
-  <PropertyGroup Label="Build">
-    <!-- Follows Reactor.Maui's TFMs (net10 only). Overridable (-p:NaluMauiReactorTargetFrameworks=net10.0)
-         so agents/CI on hosts without mobile workloads can build/test. -->
-    <NaluMauiReactorTargetFrameworksDefault>net10.0;net10.0-android;net10.0-ios26.0;net10.0-maccatalyst</NaluMauiReactorTargetFrameworksDefault>
-    <NaluMauiReactorTargetFrameworksDefault Condition="$([MSBuild]::IsOSPlatform('windows'))">$(NaluMauiReactorTargetFrameworksDefault);net10.0-windows10.0.19041.0</NaluMauiReactorTargetFrameworksDefault>
-    <NaluMauiReactorTargetFrameworks Condition="'$(NaluMauiReactorTargetFrameworks)' == ''">$(NaluMauiReactorTargetFrameworksDefault)</NaluMauiReactorTargetFrameworks>
-    <TargetFrameworks>$(NaluMauiReactorTargetFrameworks)</TargetFrameworks>
-
-    <SingleProject>true</SingleProject>
-    <OutputType>Library</OutputType>
-
-    <!-- Reactor.Maui floor: the version validated against this adapter's TemplateHost usage. -->
-    <MauiReactorVersion>4.0.18</MauiReactorVersion>
-
-    <SupportedOSPlatformVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'ios'">12.2</SupportedOSPlatformVersion>
-    <SupportedOSPlatformVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'maccatalyst'">15.0</SupportedOSPlatformVersion>
-    <SupportedOSPlatformVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'android'">21.0</SupportedOSPlatformVersion>
-    <SupportedOSPlatformVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'windows'">10.0.17763.0</SupportedOSPlatformVersion>
-    <TargetPlatformMinVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'windows'">10.0.17763.0</TargetPlatformMinVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
-      <_Parameter1>Nalu.Maui.Test</_Parameter1>
-    </AssemblyAttribute>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Maui.Controls" Version="$(MauiVersion10)" />
-    <PackageReference Include="Reactor.Maui" Version="$(MauiReactorVersion)" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Nalu.Maui.Navigation\Nalu.Maui.Navigation.csproj" />
-  </ItemGroup>
-</Project>
