+++ Tests/Nalu.Maui.Test/Nalu.Maui.Test.csproj
@@ -10,7 +10,6 @@
       <ProjectReference Include="..\..\Source\Nalu.Maui.Core\Nalu.Maui.Core.csproj" />
       <ProjectReference Include="..\..\Source\Nalu.Maui.Layouts\Nalu.Maui.Layouts.csproj" />
       <ProjectReference Include="..\..\Source\Nalu.Maui.Navigation\Nalu.Maui.Navigation.csproj" />
-      <ProjectReference Include="..\..\Source\Nalu.Maui.Navigation.MauiReactor\Nalu.Maui.Navigation.MauiReactor.csproj" />
       <ProjectReference Include="..\..\Source\Nalu.Maui.Scaffold\Nalu.Maui.Scaffold.csproj" />
       <ProjectReference Include="..\..\Source\Nalu.Maui.VirtualScroll\Nalu.Maui.VirtualScroll.csproj" />
     </ItemGroup>
@@ -38,5 +37,7 @@
         <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       </PackageReference>
       <PackageReference Include="CommunityToolkit.Mvvm" Version="8.4.0" />
+      <!-- Validates the documented MauiReactor IComponentPageFactory bridge headlessly. -->
+      <PackageReference Include="Reactor.Maui" Version="4.0.18" />
     </ItemGroup>
 </Project>
